requirements.txt
Django == 4.2.23
pymysql == 1.1.1
django-environ == 0.12.0
Pillow == 11.2.1
drf-yasg == 1.21.10
djangorestframework == 3.16.0